moved amz + bwb logic into core, fixed endpoints
