+++ openlibrary/plugins/openlibrary/api.py
@@ -11,13 +11,13 @@ from infogami.utils import delegate
 from infogami.utils.view import render_template
 from infogami.plugins.api.code import jsonapi
 from openlibrary import accounts
-from openlibrary.utils.isbn import isbn_10_to_isbn_13
+from openlibrary.utils.isbn import isbn_10_to_isbn_13, normalize_isbn
 from openlibrary.utils import extract_numeric_id_from_olid
 from openlibrary.plugins.worksearch.subjects import get_subject
 from openlibrary.core import ia, db, models, lending, cache, helpers as h
 from openlibrary.catalog.add_book import load_from_amazon_metadata
 
-from openlibrary.plugins.upstream.code import (
+from openlibrary.core.vendors import (
     get_amazon_metadata, get_betterworldbooks_metadata)
 
 class book_availability(delegate.page):
@@ -223,10 +223,13 @@ class price_api(delegate.page):
 
     @jsonapi
     def GET(self, isbn):
+        isbn = normalize_isbn(isbn)
+        isbn_type = 'isbn_' + ('13' if len(isbn) == 13 else '10')
         metadata = {
             'amazon': get_amazon_metadata(isbn) or {},
             'betterworldbooks': get_betterworldbooks_metadata(isbn) or {}
         }
+        # if bwb fails and isbn10, try again with isbn13
         if len(isbn) == 10 and \
            metadata['betterworldbooks'].get('price') is None:
             isbn_13 = isbn_10_to_isbn_13(isbn)
@@ -235,17 +238,20 @@ class price_api(delegate.page):
 
         # fetch book by isbn if it exists
         book = web.ctx.site.things({
-            "type": "/type/edition", 'isbn': isbn
+            'type': '/type/edition',
+            isbn_type: isbn,
         })
 
         # if no OL edition for isbn, attempt to create
-        if not book and metadata.get('amazon'):
+        if (not book) and metadata.get('amazon'):
             book = load_from_amazon_metadata(metadata.get('amazon'))
 
         # include ol edition metadata in response, if available
-        if book and book.get('edition'):
-            metadata['key'] = book['edition']['key']
-            if 'ocaid' in book['edition']:
-                metadata['ocaid'] = book['edition']['ocaid']
+        if book:
+            ed = web.ctx.site.get(book[0])
+            if ed:
+                metadata['key'] = ed.key
+                if getattr(ed, 'ocaid'):
+                    metadata['ocaid'] = ed.ocaid
 
         return simplejson.dumps(metadata)
